backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langgraph==0.2.0
langchain==0.3.0
langchain-core==0.3.0
httpx==0.28.0
python-dotenv==1.0.1
chromadb==0.5.0
pydantic==2.10.0
websockets==14.0
anthropic==0.42.0
